[AGW][OAI][MME] Tech Debt 2021-April: reduce compilation warnings for MME compilation (#6477)
* Tech Debt 2021-April: reduce compilation warnings for MME compilation


* fix typo


* [CI] last round


* [MME] review comment


* Fix compilation error for LI agent

+++ lte/gateway/c/oai/tasks/nas5g/include/M5GCommonDefs.h
@@ -70,7 +70,7 @@
   {                                                                            \
     int iLEN = 0;                                                              \
     if (bUFFER != NULL) {                                                      \
-      while (iLEN < lEN) {                                                     \
+      while (iLEN < ((int) lEN)) {                                             \
         MLOG(MDEBUG) << " 0x" << std::hex << int(*(bUFFER + iLEN));            \
         iLEN++;                                                                \
       }                                                                        \
